+++ src/main/java/com/backend/connectable/DataLoader.java
@@ -6,11 +6,10 @@ import com.backend.connectable.event.domain.*;
 import com.backend.connectable.event.domain.repository.EventRepository;
 import com.backend.connectable.event.domain.repository.TicketRepository;
 import com.backend.connectable.kas.service.KasService;
-import com.backend.connectable.kas.service.dto.ContractItemResponse;
+import com.backend.connectable.kas.service.contract.dto.ContractItemResponse;
 import com.backend.connectable.s3.service.S3Service;
 import com.backend.connectable.user.domain.User;
 import com.backend.connectable.user.domain.repository.UserRepository;
-import java.time.LocalDateTime;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -19,6 +18,8 @@ import org.springframework.boot.ApplicationRunner;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.time.LocalDateTime;
 
 @Profile("dev")
 @Transactional
