+++ src/main/java/com/backend/connectable/admin/service/AdminService.java
@@ -11,8 +11,8 @@ import com.backend.connectable.exception.ConnectableException;
 import com.backend.connectable.exception.ErrorType;
 import com.backend.connectable.exception.KasException;
 import com.backend.connectable.kas.service.KasService;
-import com.backend.connectable.kas.service.dto.ContractItemResponse;
-import com.backend.connectable.kas.service.dto.TransactionResponse;
+import com.backend.connectable.kas.service.common.dto.TransactionResponse;
+import com.backend.connectable.kas.service.contract.dto.ContractItemResponse;
 import com.backend.connectable.order.domain.OrderDetail;
 import com.backend.connectable.order.domain.repository.OrderDetailRepository;
 import com.backend.connectable.s3.service.S3Service;
