+++ src/main/java/com/backend/connectable/event/domain/dto/EventTicket.java
@@ -3,11 +3,12 @@ package com.backend.connectable.event.domain.dto;
 import com.backend.connectable.event.domain.TicketMetadata;
 import com.backend.connectable.event.domain.TicketSalesStatus;
 import com.querydsl.core.annotations.QueryProjection;
-import java.time.LocalDateTime;
-import javax.persistence.Convert;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import javax.persistence.Convert;
+import java.time.LocalDateTime;
 
 @Getter
 @Setter
