+++ src/main/java/com/backend/connectable/event/domain/repository/EventRepository.java
@@ -1,10 +1,11 @@
 package com.backend.connectable.event.domain.repository;
 
 import com.backend.connectable.event.domain.Event;
-import java.util.List;
-import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+
+import java.util.List;
+import java.util.Optional;
 
 public interface EventRepository extends JpaRepository<Event, Long>, EventRepositoryCustom {
