+++ src/main/java/com/backend/connectable/event/domain/repository/EventRepositoryImpl.java
@@ -1,9 +1,5 @@
 package com.backend.connectable.event.domain.repository;
 
-import static com.backend.connectable.artist.domain.QArtist.artist;
-import static com.backend.connectable.event.domain.QEvent.event;
-import static com.backend.connectable.event.domain.QTicket.ticket;
-
 import com.backend.connectable.event.domain.TicketSalesStatus;
 import com.backend.connectable.event.domain.dto.EventDetail;
 import com.backend.connectable.event.domain.dto.EventTicket;
@@ -16,10 +12,15 @@ import com.querydsl.core.types.dsl.CaseBuilder;
 import com.querydsl.core.types.dsl.NumberExpression;
 import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import org.springframework.stereotype.Repository;
+
+import javax.persistence.EntityManager;
 import java.util.List;
 import java.util.Optional;
-import javax.persistence.EntityManager;
-import org.springframework.stereotype.Repository;
+
+import static com.backend.connectable.artist.domain.QArtist.artist;
+import static com.backend.connectable.event.domain.QEvent.event;
+import static com.backend.connectable.event.domain.QTicket.ticket;
 
 @Repository
 public class EventRepositoryImpl implements EventRepositoryCustom {
