+++ src/main/java/com/backend/connectable/event/domain/repository/TicketRepository.java
@@ -1,10 +1,11 @@
 package com.backend.connectable.event.domain.repository;
 
 import com.backend.connectable.event.domain.Ticket;
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+
+import java.util.List;
 
 public interface TicketRepository extends JpaRepository<Ticket, Long>, TicketRepositoryCustom {
