+++ src/main/java/com/backend/connectable/event/ui/dto/EventDetailResponse.java
@@ -3,11 +3,12 @@ package com.backend.connectable.event.ui.dto;
 import com.backend.connectable.event.domain.SalesOption;
 import com.backend.connectable.global.common.util.DateTimeUtil;
 import com.backend.connectable.global.common.util.OpenseaCollectionNamingUtil;
-import java.time.LocalDateTime;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import java.time.LocalDateTime;
 
 @NoArgsConstructor
 @Getter
