+++ src/main/java/com/backend/connectable/event/ui/dto/TicketResponse.java
@@ -3,8 +3,9 @@ package com.backend.connectable.event.ui.dto;
 import com.backend.connectable.event.domain.TicketMetadata;
 import com.backend.connectable.event.domain.TicketSalesStatus;
 import com.backend.connectable.global.common.util.DateTimeUtil;
-import java.time.LocalDateTime;
 import lombok.*;
+
+import java.time.LocalDateTime;
 
 @Getter
 @Setter
