+++ src/main/java/com/backend/connectable/global/common/util/RedisUtil.java
@@ -1,10 +1,11 @@
 package com.backend.connectable.global.common.util;
 
-import java.time.Duration;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.stereotype.Service;
+
+import java.time.Duration;
 
 @RequiredArgsConstructor
 @Service
