+++ src/main/java/com/backend/connectable/kas/service/common/dto/TransactionResponse.java
@@ -1,4 +1,4 @@
-package com.backend.connectable.kas.service.dto;
+package com.backend.connectable.kas.service.common.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,8 +9,7 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class ContractDeployResponse {
-    private String owner;
+public class TransactionResponse {
     private String status;
     private String transactionHash;
 }
