+++ src/main/java/com/backend/connectable/kas/service/contract/dto/ContractDeployRequest.java
@@ -1,4 +1,4 @@
-package com.backend.connectable.kas.service.dto;
+package com.backend.connectable.kas.service.contract.dto;
 
 import lombok.Builder;
 import lombok.Getter;
