+++ src/main/java/com/backend/connectable/kas/service/contract/dto/ContractDeployResponse.java
@@ -1,4 +1,4 @@
-package com.backend.connectable.kas.service.dto;
+package com.backend.connectable.kas.service.contract.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,7 +9,8 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class TransactionResponse {
+public class ContractDeployResponse {
+    private String owner;
     private String status;
     private String transactionHash;
 }
