+++ src/main/java/com/backend/connectable/kas/service/token/dto/TokenHistoryResponse.java
@@ -1,4 +1,4 @@
-package com.backend.connectable.kas.service.dto;
+package com.backend.connectable.kas.service.token.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
