+++ src/main/java/com/backend/connectable/kas/service/token/dto/TokenRequest.java
@@ -1,4 +1,4 @@
-package com.backend.connectable.kas.service.dto;
+package com.backend.connectable.kas.service.token.dto;
 
 import lombok.Builder;
 import lombok.Getter;
