+++ src/main/java/com/backend/connectable/order/ui/dto/OrderRequest.java
@@ -1,13 +1,14 @@
 package com.backend.connectable.order.ui.dto;
 
 import com.backend.connectable.exception.sequence.ValidationGroups;
-import java.util.List;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Pattern;
+import java.util.List;
 
 @Getter
 @Setter
